Scroll to top when going to prev/next page removing `loading.tsx` component, using `<Suspense>` indeed
